requirements.txt
alembic
fastapi
fastapi-users[sqlalchemy]
httpx-oauth
pydantic
pyyaml
sqlalchemy[asyncio]
asyncpg
uvicorn[standard]
